Revert "refactor: move Main to testclasses - second try"
This reverts commit 47f60e73500ddcbf5ec1992c4f24c6167cb5d850.
